requirements/base.txt
dependency-injector
fastapi[all]
sqlalchemy
pydantic~=1.10
passlib[bcrypt]
python-jose[cryptography]
